backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
openai>=1.3.0
pdfly>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0
PyPDF2>=3.0.0
aiofiles>=23.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-socketio>=5.10.0
redis>=5.0.0
celery>=5.3.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
httpx>=0.25.0
websockets>=12.0